refactor: extract watcher and round-complete logic from session.go into watch.go
Move file/git watching, round-complete handlers, and their closely
related helpers (RefreshDiffs, RefreshFileList, carryForwardComment,
loadResolvedComments, carryForwardComments, emitRoundStatus) into a
dedicated watch.go file. Pure file reorganization with no behavior
changes.
